@@ -16,6 +16,16 @@ class Kernel extends ConsoleKernel
     protected function schedule(Schedule $schedule)
     {
         // $schedule->command('inspire')->hourly();
+        // $schedule->command('queue:retry all')
+        // ->everySixHours();
+
+        $schedule->command('queue:work', ['--sleep=3', '--tries=3', '--max-time=300'])
+            ->everyFiveMinutes();
+
+        $schedule->command('queue:prune-failed', ['--hours=24'])
+            ->dailyAt('01:00');
+
+        // $schedule->command('email:sendFeedback')->hourly();
     }
 
     /**
@@ -25,7 +35,7 @@ class Kernel extends ConsoleKernel
      */
     protected function commands()
     {
-        $this->load(__DIR__.'/Commands');
+        $this->load(__DIR__ . '/Commands');
 
         require base_path('routes/console.php');
     }
