@@ -11,7 +11,7 @@ use Laravel\Sanctum\HasApiTokens;
 class Admin extends Authenticatable
 {
     use HasApiTokens, HasFactory, Notifiable;
-    protected $guard='admin';
+    protected $guard = 'admin';
 
     /**
      * The attributes that are mass assignable.
@@ -42,4 +42,9 @@ class Admin extends Authenticatable
     protected $casts = [
         'email_verified_at' => 'datetime',
     ];
+
+    public function isSuper()
+    {
+        return $this->is_super;
+    }
 }
