@@ -9,8 +9,13 @@ class user_exam extends Model
 {
     use HasFactory;
 
-    protected $table="user_exams";
-    protected $primaryKey="id";
+    protected $table = 'user_exams';
+    protected $primaryKey = 'id';
 
-    protected $fillable=['user_id','exam_id','std_status','exam_joined'];
+    protected $fillable = ['user_id', 'exam_id', 'std_status', 'exam_joined', 'started', 'submitted'];
+
+    public function result()
+    {
+        return $this->hasOne(Oex_result::class, 'user_id', 'user_id');
+    }
 }
