@@ -17,9 +17,10 @@ class CreateOexExamMastersTable extends Migration
             $table->increments('id');
             $table->string('title')->nullable();
             $table->string('category')->nullable();
+            $table->string('passmark')->nullable();
             $table->string('exam_date')->nullable();
             $table->string('exam_duration')->nullable();
-            $table->string('status')->nullable(); 
+            $table->string('status')->nullable();
             $table->timestamps();
         });
     }
