@@ -16,9 +16,9 @@ class CreateOexQuestionMastersTable extends Migration
         Schema::create('oex_question_masters', function (Blueprint $table) {
             $table->id();
             $table->string('exam_id')->nullable();
-            $table->string('questions')->nullable();
+            $table->longText('questions')->nullable();
             $table->string('ans')->nullable();
-            $table->string('options')->nullable();
+            $table->longText('options')->nullable();
             $table->string('status')->nullable();
             $table->timestamps();
         });
