@@ -15,10 +15,11 @@ class UserSeeder extends Seeder
     public function run()
     {
         //
-        $user=[
-            'name'=>'student',
-            'email'=>'student@gmail.com',
-            'password'=>bcrypt('password'),
+        $user = [
+            'name' => 'student',
+            'email' => 'student@gmail.com',
+            'password' => bcrypt('password'),
+            'userId' => 'siouas879',
         ];
 
         User::create($user);
