@@ -30,7 +30,7 @@
                 <!-- Default box -->
                 <!-- /.card -->
                 <div class="card mt-4">
-                    
+
                     <div class="card-body">
                         <h2>Student information</h2>
                         <table class="table">
@@ -55,19 +55,24 @@
                                 <td>{{ $exam_info->exam_date}}</td>
                             </tr>
                         </table>
+                        <br>
                         <h2>Result info</h2>
                         <table class="table">
                             <tr>
-                                <td>Correst ans : </td>
+                                <td>Number of correct answers : </td>
                                 <td>{{ $result_info->yes_ans}}</td>
                             </tr>
                             <tr>
-                                <td>Wrong ans : </td>
+                                <td>Number of wrong answers : </td>
                                 <td>{{ $result_info->no_ans}}</td>
                             </tr>
                             <tr>
-                                <td>Total : </td>
-                                <td>{{ $result_info->yes_ans+$result_info->no_ans}}</td>
+                                <td>Total marks: </td>
+                                <td>{{ $result_info->yes_ans}}/30</td>
+                            </tr>
+                            <tr>
+                                <td>Percentage score :</td>
+                                <td>{{ round(($result_info->yes_ans / 30) * 100) }}%</td>
                             </tr>
                         </table>
                     </div>
@@ -83,5 +88,5 @@
     <!-- Modal -->
 
 
- 
+
 @endsection
